requirements.txt
defusedxml==0.4.1
Django==1.9.7
django-allauth==0.25.2
django-rest-auth==0.7.0
djangorestframework==3.3.3
oauthlib==1.1.2
python3-openid==3.0.10
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
wheel==0.24.0
